// src/lib/grpc/jobworkerp/function/data/function_set.ts
export interface FunctionSetData {
  name: string;
  description: string;
  category: number;
  targets: FunctionUsing[];
}

export interface FunctionSetId {
  value: string;
}

export interface FunctionSet {
  id: FunctionSetId | undefined;
  data: FunctionSetData | undefined;
}
// ...
export const FunctionSetData: MessageFns<FunctionSetData> = {
  encode(message: FunctionSetData, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.name !== "") {
      writer.uint32(10).string(message.name);
    }
    if (message.description !== "") {
      writer.uint32(18).string(message.description);
    }
    if (message.category !== 0) {
      writer.uint32(24).int32(message.category);
    }
    for (const v of message.targets) {
      FunctionUsing.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FunctionSetData {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFunctionSetData();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.description = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.category = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.targets.push(FunctionUsing.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<FunctionSetData>): FunctionSetData {
    return FunctionSetData.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FunctionSetData>): FunctionSetData {
    const message = createBaseFunctionSetData();
    message.name = object.name ?? "";
    message.description = object.description ?? "";
    message.category = object.category ?? 0;
    message.targets = object.targets?.map((e) => FunctionUsing.fromPartial(e)) || [];
    return message;
  },
};

function createBaseFunctionSetId(): FunctionSetId {
  return { value: "0" };
}

export const FunctionSetId: MessageFns<FunctionSetId> = {
  encode(message: FunctionSetId, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.value !== "0") {
      writer.uint32(8).int64(message.value);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FunctionSetId {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFunctionSetId();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.value = reader.int64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<FunctionSetId>): FunctionSetId {
    return FunctionSetId.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FunctionSetId>): FunctionSetId {
    const message = createBaseFunctionSetId();
    message.value = object.value ?? "0";
    return message;
  },
};

function createBaseFunctionSet(): FunctionSet {
  return { id: undefined, data: undefined };
}

export const FunctionSet: MessageFns<FunctionSet> = {
  encode(message: FunctionSet, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined) {
      FunctionSetId.encode(message.id, writer.uint32(10).fork()).join();
    }
    if (message.data !== undefined) {
      FunctionSetData.encode(message.data, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): FunctionSet {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseFunctionSet();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = FunctionSetId.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.data = FunctionSetData.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<FunctionSet>): FunctionSet {
    return FunctionSet.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<FunctionSet>): FunctionSet {
    const message = createBaseFunctionSet();
    message.id = (object.id !== undefined && object.id !== null) ? FunctionSetId.fromPartial(object.id) : undefined;
    message.data = (object.data !== undefined && object.data !== null)
      ? FunctionSetData.fromPartial(object.data)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}