// src/lib/grpc/jobworkerp/function/service/function_set.ts
export interface CreateFunctionSetResponse {
  /** generated id */
  id: FunctionSetId | undefined;
}

export interface OptionalFunctionSetResponse {
  data?: FunctionSet | undefined;
}

export interface OptionalFunctionSetDetailResponse {
  data?: FunctionSetDetail | undefined;
}

export interface FindByNameRequest {
  name: string;
}

function createBaseCreateFunctionSetResponse(): CreateFunctionSetResponse {
  return { id: undefined };
}

export const CreateFunctionSetResponse: MessageFns<CreateFunctionSetResponse> = {
  encode(message: CreateFunctionSetResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.id !== undefined) {
      FunctionSetId.encode(message.id, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CreateFunctionSetResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCreateFunctionSetResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.id = FunctionSetId.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<CreateFunctionSetResponse>): CreateFunctionSetResponse {
    return CreateFunctionSetResponse.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CreateFunctionSetResponse>): CreateFunctionSetResponse {
    const message = createBaseCreateFunctionSetResponse();
    message.id = (object.id !== undefined && object.id !== null) ? FunctionSetId.fromPartial(object.id) : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}